requirements.txt
python_dotenv
argparse
requests